Refactoring modals into useDepositModals hook
