+++ src/components/DepositWidget/index.tsx
@@ -1,40 +1,25 @@
 import React from 'react'
-import Modali, { useModali } from 'modali'
+import Modali from 'modali'
 import BN from 'bn.js'
 
 import { Row } from './Row'
 import ErrorMsg from 'components/ErrorMsg'
 import Widget from 'components/layout/Widget'
-import { ModalBodyWrapper, DepositWidgetWrapper } from './Styled'
+import { DepositWidgetWrapper } from './Styled'
 
 import { useTokenBalances } from 'hooks/useTokenBalances'
 import { useRowActions } from './useRowActions'
+import { useDepositModals } from './useDepositModals'
 import useSafeState from 'hooks/useSafeState'
 import useWindowSizes from 'hooks/useWindowSizes'
 
-import { log, formatAmount, getToken } from 'utils'
+import { log, getToken } from 'utils'
 import { ZERO } from 'const'
 import { TokenBalanceDetails } from 'types'
 
-const OverwriteModalBody: React.FC = () => {
-  return (
-    <ModalBodyWrapper>
-      <div>
-        <p>You have a pending withdraw request. </p>
-        <p>Sending this one will overwrite the pending amount.</p>
-        <p>No funds will be lost.</p>
-      </div>
-      <p>Do you wish to replace the previous withdraw request?</p>
-    </ModalBodyWrapper>
-  )
-}
-
-const WithdrawAndClaimModalBody: React.FC = () => {
-  return (
-    <ModalBodyWrapper>
-      <p>By sending this withdraw request, you will automatically receive all claimable amounts back to your wallet.</p>
-    </ModalBodyWrapper>
-  )
+interface WithdrawState {
+  amount: BN
+  tokenAddress: string
 }
 
 const DepositWidget: React.FC = () => {
@@ -42,73 +27,31 @@ const DepositWidget: React.FC = () => {
   const { enableToken, deposit, requestWithdraw, claim } = useRowActions({ balances, setBalances })
   const windowSpecs = useWindowSizes()
 
-  const [withdrawRequest, setWithdrawRequest] = useSafeState({
+  const [withdrawRequest, setWithdrawRequest] = useSafeState<WithdrawState>({
     amount: ZERO,
     tokenAddress: '',
-    pendingAmount: '',
-    symbol: '',
   })
 
-  const [withdrawOverwriteModal, toggleWithdrawOverwriteModal] = useModali({
-    centered: true,
-    animated: true,
-    title: 'Confirm withdraw overwrite',
-    message: <OverwriteModalBody />,
-    buttons: [
-      <Modali.Button label="Cancel" key="no" isStyleCancel onClick={(): void => toggleWithdrawOverwriteModal()} />,
-      <Modali.Button
-        label="Continue"
-        key="yes"
-        isStyleDefault
-        onClick={async (): Promise<void> => {
-          // On confirm, do the request
-          toggleWithdrawOverwriteModal()
-          await requestWithdraw(withdrawRequest.amount, withdrawRequest.tokenAddress)
-        }}
-      />,
-    ],
-  })
-
-  const [withdrawAndClaimModal, toggleWithdrawAndClaimModal] = useModali({
-    centered: true,
-    animated: true,
-    title: 'Please note',
-    message: <WithdrawAndClaimModalBody />,
-    buttons: [
-      <Modali.Button label="Cancel" key="no" isStyleCancel onClick={(): void => toggleWithdrawAndClaimModal()} />,
-      <Modali.Button
-        label="Continue"
-        key="yes"
-        isStyleDefault
-        onClick={async (): Promise<void> => {
-          // On confirm, do the request
-          toggleWithdrawAndClaimModal()
-          await requestWithdraw(withdrawRequest.amount, withdrawRequest.tokenAddress)
-        }}
-      />,
-    ],
-  })
+  const {
+    withdrawOverwriteModal,
+    toggleWithdrawOverwriteModal,
+    withdrawAndClaimModal,
+    toggleWithdrawAndClaimModal,
+  } = useDepositModals({ ...withdrawRequest, requestWithdraw })
 
   const requestWithdrawConfirmation = async (amount: BN, tokenAddress: string, claimable: boolean): Promise<void> => {
-    const { withdrawingBalance, decimals, symbol } = getToken('address', tokenAddress, balances) as Required<
-      TokenBalanceDetails
-    >
+    const { withdrawingBalance, symbol } = getToken('address', tokenAddress, balances) as Required<TokenBalanceDetails>
+
     log(`Confirm withdrawal for ${symbol} with withdrawingBalance ${withdrawingBalance}`)
+
     if (!withdrawingBalance.isZero()) {
       // Storing current values before displaying modal
       setWithdrawRequest({
         amount,
         tokenAddress,
-        pendingAmount: formatAmount(withdrawingBalance, decimals) || '',
-        symbol,
       })
 
-      if (claimable) {
-        toggleWithdrawAndClaimModal()
-      } else {
-        // Confirm withdrawal: There's an unclaimed withdraw request
-        toggleWithdrawOverwriteModal()
-      }
+      claimable ? toggleWithdrawAndClaimModal() : toggleWithdrawOverwriteModal()
     } else {
       // No need to confirm the withdrawal: No amount is pending to be claimed
       await requestWithdraw(amount, tokenAddress)
